app/requirements.txt
pandas
openai
tiktoken
python-dotenv